+++ spec/rubocop/cop/rails/match_route_spec.rb
@@ -63,6 +63,21 @@ RSpec.describe RuboCop::Cop::Rails::MatchRoute do
     RUBY
   end
 
+  it 'registers an offense when using match with string interpolation' do
+    expect_offense(<<~'RUBY')
+      routes.draw do
+        match "#{resource}/:action/:id", via: [:put]
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `put` instead of `match` to define a route.
+      end
+    RUBY
+
+    expect_correction(<<~'RUBY')
+      routes.draw do
+        put "#{resource}/:action/:id"
+      end
+    RUBY
+  end
+
   it 'does not register an offense when not within routes' do
     expect_no_offenses(<<~RUBY)
       match 'photos/:id', to: 'photos#show', via: :get
@@ -89,6 +104,14 @@ RSpec.describe RuboCop::Cop::Rails::MatchRoute do
     expect_no_offenses(<<~RUBY)
       routes.draw do
         get 'photos/:id', to: 'photos#show'
+      end
+    RUBY
+  end
+
+  it 'does not register an offense when via is a variable' do
+    expect_no_offenses(<<~'RUBY')
+      routes.draw do
+        match ':controller/:action/:id', via: method
       end
     RUBY
   end